changed the value the id of the employee so it compares the longvalue

+++ service/src/main/java/com/globati/service/EmployeeService.java
@@ -298,7 +298,7 @@ public class EmployeeService {
         try {
 
             //It is ok for to update for one of us to update our profiles to a reserved word name.
-            if( userNameIsAReservedWord(employee.getGlobatiUsername()) && (employee.getId() != 1 && employee.getId() != 17 && employee.getId() != 49) ){
+            if( userNameIsAReservedWord(employee.getGlobatiUsername()) && (employee.getId().longValue() != 1L && employee.getId().longValue() != 17L && employee.getId().longValue() != 49L) ){
                 throw new UserDoesNotExistException("Username is a reserved word for user: "+employee.getGlobatiUsername());
             }
             return this.employeeRepository.save(employee);
